+++ R/RcppExports.R
@@ -129,6 +129,10 @@ CPL_geos_union <- function(sfc, by_feature = FALSE) {
     .Call('_sf_CPL_geos_union', PACKAGE = 'sf', sfc, by_feature)
 }
 
+CPL_geos_snap <- function(sfc0, sfc1, tolerance) {
+    .Call('_sf_CPL_geos_snap', PACKAGE = 'sf', sfc0, sfc1, tolerance)
+}
+
 CPL_geos_op <- function(op, sfc, bufferDist, nQuadSegs = 30L, dTolerance = 0.0, preserveTopology = FALSE, bOnlyEdges = 1L, dfMaxLength = 0.0) {
     .Call('_sf_CPL_geos_op', PACKAGE = 'sf', op, sfc, bufferDist, nQuadSegs, dTolerance, preserveTopology, bOnlyEdges, dfMaxLength)
 }